# NFC: rewrite EnumSet to handle larger enums.