--- a/source/enum_set.h
+++ b/source/enum_set.h
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <stdint.h>
-
+#include <cstdint>
 #include <functional>
 #include <initializer_list>
 #include <limits>