--- a/source/enum_set.h
+++ b/source/enum_set.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2016 Google Inc.
+// Copyright (c) 2023 Google Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,192 +12,207 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <stdint.h>
+
+#include <algorithm>
+#include <functional>
+#include <initializer_list>
+#include <limits>
+#include <type_traits>
+#include <vector>
+
 #ifndef SOURCE_ENUM_SET_H_
 #define SOURCE_ENUM_SET_H_
 
-#include <cstdint>
-#include <functional>
-#include <memory>
-#include <set>
-#include <utility>
-
 #include "source/latest_version_spirv_header.h"
-#include "source/util/make_unique.h"
 
 namespace spvtools {
 
-// A set of values of a 32-bit enum type.
-// It is fast and compact for the common case, where enum values
-// are at most 63.  But it can represent enums with larger values,
-// as may appear in extensions.
-template <typename EnumType>
+template <typename T>
 class EnumSet {
- private:
-  // The ForEach method will call the functor on enum values in
-  // enum value order (lowest to highest).  To make that easier, use
-  // an ordered set for the overflow values.
-  using OverflowSetType = std::set<uint32_t>;
+  using BucketType = uint64_t;
+  using ElementType = std::underlying_type_t<T>;
+  static_assert(std::is_enum_v<T>, "EnumSets only works with enums.");
+  static_assert(std::is_signed_v<ElementType> == false,
+                "EnumSet doesn't supports signed enums.");
+  static_assert(sizeof(T) * 8ULL <= std::numeric_limits<ElementType>::max());
+  static constexpr size_t BUCKET_SIZE = sizeof(BucketType) * 8ULL;
+
+  struct Bucket {
+    BucketType data;
+    T start;
+  };
+
+  std::vector<Bucket> buckets;
+
+  static constexpr inline size_t compute_bucket_index(T value) {
+    return static_cast<size_t>(value) / BUCKET_SIZE;
+  }
+
+  static constexpr inline size_t compute_bucket_offset(T value) {
+    return static_cast<ElementType>(value) % BUCKET_SIZE;
+  }
+
+  static constexpr inline T compute_bucket_start(T value) {
+    return static_cast<T>(BUCKET_SIZE * compute_bucket_index(value));
+  }
+
+  static constexpr inline uint64_t compute_mask_for_value(T value) {
+    return 1ULL << compute_bucket_offset(value);
+  }
+
+  static constexpr inline T get_value_from_bucket(const Bucket& bucket,
+                                                  ElementType offset) {
+    return static_cast<T>(static_cast<ElementType>(bucket.start) + offset);
+  }
+
+  size_t find_bucket_for_value(T value) const {
+    // Set is empty, insert at 0.
+    if (buckets.size() == 0) {
+      return 0;
+    }
+
+    size_t index = std::min(buckets.size() - 1, compute_bucket_index(value));
+    const T needle = compute_bucket_start(value);
+
+    const T bucket_start = buckets[index].start;
+    // Computed index is the correct one.
+    if (bucket_start == needle) {
+      return index;
+    }
+
+    // Bucket contains smaller values. Linear scan right.
+    if (bucket_start < needle) {
+      for (index += 1; index < buckets.size() && buckets[index].start < needle;
+           index++) {
+      }
+      return index;
+    }
+
+    // Bucket contains larger values, insert front.
+    if (index == 0) {
+      return index;
+    }
+
+    for (index -= 1; index > 0 && buckets[index].start > needle; index--) {
+    }
+    return buckets[index].start >= needle ? index : index + 1;
+  }
+
+  void create_bucket_for(size_t index, T value) {
+    Bucket bucket = {1ULL << compute_bucket_offset(value),
+                     compute_bucket_start(value)};
+    buckets.emplace(buckets.begin() + index, std::move(bucket));
+  }
 
  public:
-  // Construct an empty set.
-  EnumSet() {}
-  // Construct an set with just the given enum value.
-  explicit EnumSet(EnumType c) { Add(c); }
-  // Construct an set from an initializer list of enum values.
-  EnumSet(std::initializer_list<EnumType> cs) {
-    for (auto c : cs) Add(c);
-  }
-  EnumSet(uint32_t count, const EnumType* ptr) {
-    for (uint32_t i = 0; i < count; ++i) Add(ptr[i]);
-  }
-  // Copy constructor.
-  EnumSet(const EnumSet& other) { *this = other; }
-  // Move constructor.  The moved-from set is emptied.
-  EnumSet(EnumSet&& other) {
-    mask_ = other.mask_;
-    overflow_ = std::move(other.overflow_);
-    other.mask_ = 0;
-    other.overflow_.reset(nullptr);
-  }
-  // Assignment operator.
-  EnumSet& operator=(const EnumSet& other) {
-    if (&other != this) {
-      mask_ = other.mask_;
-      overflow_.reset(other.overflow_ ? new OverflowSetType(*other.overflow_)
-                                      : nullptr);
+  EnumSet() : buckets(0) {}
+
+  EnumSet(T value) : EnumSet() { Add(value); }
+
+  EnumSet(std::initializer_list<T> data) : EnumSet() {
+    for (auto item : data) {
+      Add(item);
     }
+  }
+
+  EnumSet(uint32_t count, const T* values) : EnumSet() {
+    for (uint32_t i = 0; i < count; i++) {
+      Add(values[i]);
+    }
+  }
+
+  EnumSet(const EnumSet& other) : buckets(other.buckets) {}
+
+  EnumSet(EnumSet&& other) : buckets(std::move(other.buckets)) {}
+
+  EnumSet& operator=(const EnumSet& other) {
+    buckets = other.buckets;
     return *this;
   }
 
-  friend bool operator==(const EnumSet& a, const EnumSet& b) {
-    if (a.mask_ != b.mask_) {
+  void Add(T value) {
+    const size_t index = find_bucket_for_value(value);
+    if (index >= buckets.size() ||
+        buckets[index].start != compute_bucket_start(value)) {
+      create_bucket_for(index, value);
+      return;
+    }
+    auto& bucket = buckets[index];
+    bucket.data |= compute_mask_for_value(value);
+  }
+
+  void Remove(T value) {
+    const size_t index = find_bucket_for_value(value);
+    if (index >= buckets.size() ||
+        buckets[index].start != compute_bucket_start(value)) {
+      return;
+    }
+    auto& bucket = buckets[index];
+    bucket.data &= ~compute_mask_for_value(value);
+    if (bucket.data == 0) {
+      buckets.erase(buckets.cbegin() + index);
+    }
+  }
+
+  bool Contains(T value) const {
+    const size_t index = find_bucket_for_value(value);
+    if (index >= buckets.size() ||
+        buckets[index].start != compute_bucket_start(value)) {
+      return false;
+    }
+    auto& bucket = buckets[index];
+    return bucket.data & compute_mask_for_value(value);
+  }
+
+  void ForEach(std::function<void(T)> f) const {
+    for (const auto& bucket : buckets) {
+      for (uint8_t i = 0; i < BUCKET_SIZE; i++) {
+        if (bucket.data & (1ULL << i)) {
+          f(get_value_from_bucket(bucket, i));
+        }
+      }
+    }
+  }
+
+  bool IsEmpty() const { return buckets.size() == 0; }
+
+  friend bool operator==(const EnumSet& lhs, const EnumSet& rhs) {
+    if (lhs.buckets.size() != rhs.buckets.size()) {
       return false;
     }
 
-    if (a.overflow_ == nullptr && b.overflow_ == nullptr) {
-      return true;
+    for (size_t i = 0; i < lhs.buckets.size(); i++) {
+      if (rhs.buckets[i].start != lhs.buckets[i].start ||
+          rhs.buckets[i].data != lhs.buckets[i].data) {
+        return false;
+      }
     }
-
-    if (a.overflow_ == nullptr || b.overflow_ == nullptr) {
-      return false;
-    }
-
-    return *a.overflow_ == *b.overflow_;
-  }
-
-  friend bool operator!=(const EnumSet& a, const EnumSet& b) {
-    return !(a == b);
-  }
-
-  // Adds the given enum value to the set.  This has no effect if the
-  // enum value is already in the set.
-  void Add(EnumType c) { AddWord(ToWord(c)); }
-
-  // Removes the given enum value from the set.  This has no effect if the
-  // enum value is not in the set.
-  void Remove(EnumType c) { RemoveWord(ToWord(c)); }
-
-  // Returns true if this enum value is in the set.
-  bool Contains(EnumType c) const { return ContainsWord(ToWord(c)); }
-
-  // Applies f to each enum in the set, in order from smallest enum
-  // value to largest.
-  void ForEach(std::function<void(EnumType)> f) const {
-    for (uint32_t i = 0; i < 64; ++i) {
-      if (mask_ & AsMask(i)) f(static_cast<EnumType>(i));
-    }
-    if (overflow_) {
-      for (uint32_t c : *overflow_) f(static_cast<EnumType>(c));
-    }
-  }
-
-  // Returns true if the set is empty.
-  bool IsEmpty() const {
-    if (mask_) return false;
-    if (overflow_ && !overflow_->empty()) return false;
     return true;
   }
 
-  // Returns true if the set contains ANY of the elements of |in_set|,
-  // or if |in_set| is empty.
-  bool HasAnyOf(const EnumSet<EnumType>& in_set) const {
-    if (in_set.IsEmpty()) return true;
+  friend bool operator!=(const EnumSet& lhs, const EnumSet& rhs) {
+    return !(lhs == rhs);
+  }
 
-    if (mask_ & in_set.mask_) return true;
-
-    if (!overflow_ || !in_set.overflow_) return false;
-
-    for (uint32_t item : *in_set.overflow_) {
-      if (overflow_->find(item) != overflow_->end()) return true;
+  bool HasAnyOf(const EnumSet<T>& in_set) const {
+    if (in_set.IsEmpty()) {
+      return true;
     }
 
+    for (auto& lhs_bucket : buckets) {
+      for (auto& rhs_bucket : in_set.buckets) {
+        if (lhs_bucket.start != rhs_bucket.start) {
+          continue;
+        }
+
+        if (lhs_bucket.data & rhs_bucket.data) {
+          return true;
+        }
+      }
+    }
     return false;
   }
-
- private:
-  // Adds the given enum value (as a 32-bit word) to the set.  This has no
-  // effect if the enum value is already in the set.
-  void AddWord(uint32_t word) {
-    if (auto new_bits = AsMask(word)) {
-      mask_ |= new_bits;
-    } else {
-      Overflow().insert(word);
-    }
-  }
-
-  // Removes the given enum value (as a 32-bit word) from the set.  This has no
-  // effect if the enum value is not in the set.
-  void RemoveWord(uint32_t word) {
-    if (auto new_bits = AsMask(word)) {
-      mask_ &= ~new_bits;
-    } else {
-      auto itr = Overflow().find(word);
-      if (itr != Overflow().end()) Overflow().erase(itr);
-    }
-  }
-
-  // Returns true if the enum represented as a 32-bit word is in the set.
-  bool ContainsWord(uint32_t word) const {
-    // We shouldn't call Overflow() since this is a const method.
-    if (auto bits = AsMask(word)) {
-      return (mask_ & bits) != 0;
-    } else if (auto overflow = overflow_.get()) {
-      return overflow->find(word) != overflow->end();
-    }
-    // The word is large, but the set doesn't have large members, so
-    // it doesn't have an overflow set.
-    return false;
-  }
-
-  // Returns the enum value as a uint32_t.
-  uint32_t ToWord(EnumType value) const {
-    static_assert(sizeof(EnumType) <= sizeof(uint32_t),
-                  "EnumType must statically castable to uint32_t");
-    return static_cast<uint32_t>(value);
-  }
-
-  // Determines whether the given enum value can be represented
-  // as a bit in a uint64_t mask. If so, then returns that mask bit.
-  // Otherwise, returns 0.
-  uint64_t AsMask(uint32_t word) const {
-    if (word > 63) return 0;
-    return uint64_t(1) << word;
-  }
-
-  // Ensures that overflow_set_ references a set.  A new empty set is
-  // allocated if one doesn't exist yet.  Returns overflow_set_.
-  OverflowSetType& Overflow() {
-    if (overflow_.get() == nullptr) {
-      overflow_ = MakeUnique<OverflowSetType>();
-    }
-    return *overflow_;
-  }
-
-  // Enums with values up to 63 are stored as bits in this mask.
-  uint64_t mask_ = 0;
-  // Enums with values larger than 63 are stored in this set.
-  // This set should normally be empty or very small.
-  std::unique_ptr<OverflowSetType> overflow_ = {};
 };
 
 // A set of spv::Capability, optimized for small capability values.