@Keenuts 2023-06-23
`CI-ndk-build` fails because it still builds in C++11 (see #5290)