@Keenuts 2023-07-06
As discussed offline:
Replacing the find function with a simple std::lower_bound does impact performance compared to this method due to the data/patterns we have.
(6% to 60% impact depending on the set size)